Fixing error during sign-up verification code process
SignUp.js was modified to fetch the user's email from the Redux store. This is bc once we submit the sign in details, a boolean flag gets toggled which will make the UI render the prompt for the verification code.

Before this modification, the username provided to confirm signin using AWS Auth API was undefined due to the boolean logic explained above.

When the verification prompt is rendered, the form with the sign in details is essentially unmounted. So we cannot pull email from the "data" property from React-hook-form