Refactored forgot password logic.
The original idea was to create a forgot password link component.
This component would have a link that navigates to the form.
Instead of adding an entirely new route for the FP form,
We can add some state in the sign in HOC to toggle back and forth between:
  - The sign in form
  - The forgot password form

The async nature of the functionality makes it a much more seamless exp for the end user.